MM-43187 Move most internal stuff out of types package
