+++ components/admin_console/team_channel_settings/channel/details/channel_moderation.tsx
@@ -9,7 +9,6 @@ import classNames from 'classnames';
 
 import {ChannelModeration as ChannelPermissions} from 'mattermost-redux/types/channels';
 import {Permissions, Roles} from 'mattermost-redux/constants';
-import {ChannelModerationRoles} from 'mattermost-redux/types/roles';
 
 import FormattedMarkdownMessage from 'components/formatted_markdown_message';
 
@@ -17,6 +16,8 @@ import {t} from 'utils/i18n';
 
 import AdminPanel from 'components/widgets/admin_console/admin_panel';
 import CheckboxCheckedIcon from 'components/widgets/icons/checkbox_checked_icon';
+
+import {ChannelModerationRoles} from './types';
 
 const PERIOD_TO_SLASH_REGEX = /\./g;
