+++ components/admin_console/team_channel_settings/channel/details/types.ts
@@ -1,8 +1,4 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-export type Typing = {
-    [x: string]: {
-        [x: string]: number;
-    };
-};
+export type ChannelModerationRoles = 'members' | 'guests';
