+++ packages/mattermost-redux/src/types/websocket.ts
@@ -1,4 +1,22 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-export * from '@mattermost/types/websocket';
+export type WebSocketBroadcast = {
+    omit_users: Record<string, boolean>;
+    user_id: string;
+    channel_id: string;
+    team_id: string;
+}
+
+export type WebSocketMessage<T> = {
+    event: string;
+    data: T;
+    broadcast: WebSocketBroadcast;
+    seq: number;
+}
+
+export type WebsocketStatus = {
+    connected: boolean;
+    lastConnectAt: number;
+    lastDisconnectAt: number;
+}
