+++ packages/types/src/client4.ts
@@ -14,24 +14,6 @@ export type ClientResponse<T> = {
     data: T;
 };
 
-type ErrorOffline = {
-    message: string;
-    url: string;
-};
-type ErrorInvalidResponse = {
-    intl: {
-        id: string;
-        defaultMessage: string;
-    };
-};
-export type ErrorApi = {
-    message: string;
-    server_error_id: string;
-    status_code: number;
-    url: string;
-};
-export type Client4Error = ErrorOffline | ErrorInvalidResponse | ErrorApi;
-
 export type Options = {
     headers?: { [x: string]: string };
     method?: string;
