+++ packages/types/src/general.ts
@@ -20,23 +20,3 @@ export type SystemSetting = {
     name: string;
     value: string;
 };
-
-export enum LicenseSkus {
-    E10 = 'E10',
-    E20 = 'E20',
-    Starter = 'starter',
-    Professional = 'professional',
-    Enterprise = 'enterprise',
-}
-
-export const isEnterpriseLicense = (license?: ClientLicense) => {
-    switch (license?.SkuShortName) {
-    case LicenseSkus.Enterprise:
-    case LicenseSkus.E20:
-        return true;
-    }
-
-    return false;
-};
-
-export const isNonEnterpriseLicense = (license?: ClientLicense) => !isEnterpriseLicense(license);
