+++ packages/types/src/insights.ts
@@ -1,5 +1,6 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+
 export enum InsightsWidgetTypes {
     TOP_CHANNELS = 'TOP_CHANNELS',
     TOP_REACTIONS = 'TOP_REACTIONS',
