+++ packages/types/src/integrations.ts
@@ -113,22 +113,6 @@ export type IntegrationsState = {
     commands: IDMappedObjects<Command>;
 };
 
-export type InteractiveDialogConfig = {
-    app_id: string;
-    trigger_id: string;
-    url: string;
-    dialog: {
-        callback_id: string;
-        title: string;
-        introduction_text: string;
-        icon_url?: string;
-        elements: DialogElement[];
-        submit_label: string;
-        notify_on_cancel: boolean;
-        state: string;
-    };
-};
-
 export type DialogSubmission = {
     url: string;
     callback_id: string;
