+++ packages/types/src/roles.ts
@@ -1,6 +1,5 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
-export type ChannelModerationRoles = 'members' | 'guests';
 
 export type Role = {
     id: string;
