+++ packages/types/src/schemes.ts
@@ -1,5 +1,6 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+
 export type SchemeScope = 'team' | 'channel';
 export type Scheme = {
     id: string;
